Add fall-through comments
GCC warns about switch statement fall-throughs unless they are annotated
with a /* fall through */ comment. Add one of these.

As for the other comment, GCC also has a slightly less strict mode that
understands the comment as long as it has the words "fall through"
somewhere in it. If we slightly tweak the wording of this comment, we can
use that mode to avoid the warning.