Initialize kinds array
GCC warns that ka[1] is possibly being used without being initialized. As
far as I can tell from reading the loop, a1 > 0 here, and the elements of
ka are initialized up to a1-1, and ka[1] is never accessed if a1 == 1.

Nonetheless, rather than rewriting the loop so GCC can understand what's
going on, maybe it's better just to put an initializer here.