Allow setting extra flags to "make" in Test Makes
When Make detects that it is running under another Make process, it prints
out "Entering directory ..." / "Leaving directory ..." messages which find
their way into the test log, and mess up the output. Normally the -s flag
is supposed to silence these messages, but it seems that GNU Make requires
an extra flag --no-print-directory here.

This sets up a global intest variable for Make flags, that is usually just
-s, but on Linux the extra flag is added.