Add Flatpak integration target to indoc instructions
I found that when building a Flatpak package where the build of each
component is siloed from the others, I need to copy the integration files
into some shared location. Ideally this would be governed by a setting in
make-integration-settings.mk, but for now, add a separate target to indoc
which I will select with INDOCOPTS in my integration settings.