Bless output of indoc tests for Windows app
The ideal output didn't reflect the recent changes to the JavaScript paste
code. I'm assuming that the current output is correct as-is.

Maybe it makes less sense to have separate tests for Windows and Mac now,
since the HTML is now identical. But the CSS still differs, so I've kept
them for now.