Alias PLATFORM_LINUX to PLATFORM_UNIX in retrospective sources
PLATFORM_LINUX is new in Inform 10.1. In the retrospective sources,
defining PLATFORM_LINUX does nothing, meaning that in the absence of a
platform definition, PLATFORM_MACOSX will take effect. Add a check for
PLATFORM_LINUX to the start of each of the ni.c sources, and treat it the
same as PLATFORM_UNIX.

(These files are generated and technically not supposed to be edited, but
the source code for old versions of Inform is now frozen, so they are not
going to be generated again.)