Modifications to support the advice pages on Linux
The Linux app doesn't currently have a launcher window that can display
these pages, but I'm working on one. In advance of that, here are Linux
versions of the advice pages.

I also found that 'make integration' didn't install the welcome banner
for the launcher window, nor one of the assets needed for the advice
pages, so this also patches the make script to do so.