Robustness for parallel builds
When the environment variable MAKEFLAGS contains -j or similar parallel
building flag, a clean build (scripts/first.sh from a fresh checkout)
will lead to errors such as

    (Step 3 of 4: building the tools)
    make -f inblorb/inblorb.mk force
    inbuild/Tangled/inbuild -rebuild -contents-of
    inform7/Internal/Extensions/Graham\ Nelson/Standard\
    Rules.i7xd/Materials/Inter
    make: inbuild/Tangled/inbuild: No such file or directory
    make: *** No rule to make target 'inform7/Tangled/Syntax.preform',
    needed by 'inform7/Internal/Extensions/Graham Nelson/English
    Language.i7xd/Materials/Languages/English/Syntax.preform'.  Stop.
    make: *** Waiting for unfinished jobs....
    make: *** [makefile:331: forcekits] Error 127

and others. This is due to the 'force' build trying to build forcetools,
forcekits and localintegration in parallel, even though there is
actually a strict dependency between them that is not fully described by
the makefile rules.

This patch introduces two changes to avoid this issue.

First of all, we introduce the .WAIT dependency before localintegration
in the 'all' and 'force' targets. This implicitly-defined phony
dependency is used by GNU Make to indicate that the dependencies on the
right must wait for the dependencies on the left to finish.

Secondly, the kits and forcekits targets now depend on the tools used to
build the kits (i.e. Inbuild and inter). For the forcekits target, this
requires the introduction of force_* variants for the tools, that were
previously built using a single target.

As a bonus, this allows the tools themselves to now be built in parallel
even during a force build.

Similar changes can be implemented for other targets that currently
build multiple files serially, but this will be left to later commits.